Log node boot progress
This is intended to help us debug errors during node boots. The two
particular cases are when a node appears to be half awake responding to
RPC requests while the HTTP interfaces aren't live.

The second is a developer friendliness issue when a node fails to boot
with the opaque "init terminating in do_boot" error.

BugzId: 28716